requirements.txt
pandas>=2.2
numpy>=1.26
matplotlib>=3.8
seaborn>=0.13
scikit-learn>=1.5
python-dotenv>=1.0
google-cloud-bigquery>=3.25
db-dtypes>=1.3
pandas-gbq>=0.23
openpyxl>=3.1
